last_time

+++ ft_strdup.c
@@ -1,13 +1,25 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strdup.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: satabay <satabay@student.42istanbul.com    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/06/25 10:11:06 by satabay           #+#    #+#             */
+/*   Updated: 2025/06/25 10:38:43 by satabay          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-char*	ft_strdup(const char *s)
+char	*ft_strdup(const char *s)
 {
-	int		i;
+	size_t	i;
 	size_t	len;
 	char	*copy;
 
 	len = ft_strlen(s);
-	copy = malloc(sizeof(char)*(len + 1));
+	copy = malloc(sizeof(char) * (len + 1));
 	if (!copy)
 		return (0);
 	i = 0;
@@ -17,5 +29,5 @@ char*	ft_strdup(const char *s)
 		i++;
 	}
 	copy[i] = '\0';
-	return (copy);	
+	return (copy);
 }
